Fixed strange null error for the fileUri.
For some reason the fileUri was null in the onResult method even though
it was being set before starting an activity. I could not figure out
what would cause this behavior so I just saved it in sharedpreferences
as a workaround.

However it still has issues.
If the image returned by the camera intent is too big it causes the
image to not be shown. The limit is 4096X2000something I believe.
To fix this the bitmap could be scaled down to a level that is known to
be reasonable whilst keeping the same aspect ratio.